pyproject.toml
[tool.poetry]
name = "tailorhustle"
version = "0.1.0"
description = "SAS for Tailors and Brands to list and provide their clothing facilities online."
authors = ["Taimoor , Awais"]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.4.1"
Django = "3.2.12"
pytz = "2021.3"
sqlparse = "0.4.2"
psycopg2-binary = "2.9.6"
Pillow = "^9.0.1"
django-crispy-forms = "1.13.0"
django-cleanup = "5.2.0"
install = "^1.3.5"
django-chartjs = "^2.3.0"
django-pwa = "^1.0.10"
timeago = "^1.0.15"
djangorestframework = "^3.13.1"
"pushbullet.py" = "^0.12.0"
django-extensions = "^3.2.0"
gunicorn = "^20.1.0"
whitenoise = "^6.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
